refactor: モード選択ボタン群を共通化
